+++ sources/reader/methods/br_tell.c
@@ -6,13 +6,13 @@
 /*   By: mgama <mgama@student.42lyon.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/21 22:24:41 by mgama             #+#    #+#             */
-/*   Updated: 2024/04/23 12:15:29 by mgama            ###   ########.fr       */
+/*   Updated: 2024/07/02 19:15:33 by mgama            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "reader/binary_reader.h"
 
-uint32_t	br_tell(t_binary_reader *this)
+uint64_t	br_tell(t_binary_reader *this)
 {
 	return (this->_pos);
 }
